requirements.txt
chardet
requests
tqdm
ftfy
regex
torch
torchvision
absl-py